ocr-backend/requirements.txt
flask
flask-cors
pillow
pix2tex
